backend/services/api_service/api/serializers.py
from django.contrib.auth.models import Group, User
from rest_framework import serializers

class UserSerializer(serializers.HyperlinkedModelSerializer):
		
	class Meta:
		model = User
		fields = ['url', 'username', 'email', 'is_staff']

class GroupSerializer(serializers.HyperlinkedModelSerializer):

	class Meta:
		model = Group
		fields = ['url', 'name']