backend/services/api_service/requirements.txt
Django==5.1.1
psycopg
djangorestframework
fastapi 
uvicorn[standard]